requirements.txt
streamlit
langchain
langchain-community
langchain-groq
chromadb
python-dotenv
requests
sentence-transformers
